Use a redigo connection pool instead of a generic resource pool
The resource pool used didn’t support returning closed connections.
Doing so would result in “use of closed network connection” errors on
subsequent usages of returned connections. On the other hand, not
returning a closed connection would exhaust the pool and lock the poller
and/or workers.

This commit drops all indirections to redigo connections and pools and
thus gracefully handles intermittent network failures and restores.